Refactor to remove manual unrolling of packet start.
Summary:
Change made to pass UNH test 4.1.2 (bugzilla 16174),
which tests that frame fragments and runts do not
cause the next valid to be dropped.

The issue with unrolling the first few words of
input is that the data valid can go low on a nibble
boundary, causing the input to block. The timing
improvements from moving to unsafe pointers has
meant that the unrolling can be removed and the
buffer logic incorporated into the select statement.

Note: It requires fschedule to be re-enabled to meet
timing due to FNOPS in the critical path.

Reviewers: davel, peter

Reviewed By: peter

Differential Revision: http://phabricator.xmos.local:9000/D72