Fixes issue with missing packet_read() notifications.
Summary:
Missing default case in MII lite server:
Was blocking on the waiteu and not executing the
filtering code outside the select {}. This meant
that packets were missed when stressed with
minimum interframe gaps.

RT MII server:
If the MAC notifies the client that a packet is ready,
and another packetarrives before the client has processed
this event, a notification for the new packet will
be issued but lost as there is no buffering of notifications.

The solution here is to renotify the client when the
copy of the first packet is complete, and continue
to notify on completion as long as there are packets
queued in the client buffer.

Adds braces to if statements for consistency

Reviewers: davel, peter

Reviewed By: peter

Differential Revision: http://phabricator.xmos.local:9000/D78