Make MII layer API public
Summary:
This commit extracts the MII layer API and makes it public. It then uses
this for the mii_ethernet compnent. This way other libraries (e.g. the
TCP/IP library) can interface directly with MII, bypassing the MAC for
reduced resource usage.

At this moment, the loopback demo is untested.

This commit also fixes a couple of issues in the standard ethernet layer
todo with the "next valid bank" fifo.

Reviewers: andy, peter

Reviewed By: peter

Differential Revision: http://phabricator.xmos.local:9000/D120