Re-implement the standard MII MAC buffering.
Summary:
The MII MAC buffering is now merged into a single buffer so that packets
can be put into the high or low priority queues depending on the output
of the filtering stage.

There is now one big buffer and a queue of packets for high and low
priority traffic for the TX. The RX has the same with an additional
queue for packets that have not yet been filtered yet.

Merge

Fixed bugs introduced when adding the mii interface.

Update test applications to use new mii interface.

This allows tests to receive line rate packets using the lite MAC.

Reviewers: andy, davel

Reviewed By: andy, davel

Differential Revision: http://phabricator.xmos.local:9000/D122