Get RT MAC to correctly cope with high/low priority packets.
Summary:
Adds a threshold for when the low priority packets get dropped.
Also fixes a number of bugs in the RT buffering.

Added filtering to RGMII.

This involves updating a lot of tests to support both high and low
priority versions.

There is also an additional test purely for the traffic queues which has
the ability to set rate limits as the RGMII MAC can't support full line
rate all the way through to the client application.

Reviewers: davel, andy

Differential Revision: http://phabricator.xmos.local:9000/D126